move mock delegate to file
